RAG_QA/LangChain RAG/requirements.txt
langchain==0.1.17 
langchain-community==0.0.36 
langchain-openai==0.0.7 
langchain-text-splitters==0.0.1 
faiss-cpu 
pypdf 
openai 
tiktoken 
python-dotenv